Refactor command result in client utils to reduce package cycles.
