+++ project-set/components/cli-utils/src/main/java/org/openrepose/cli/command/Command.java
@@ -1,6 +1,6 @@
 package org.openrepose.cli.command;
 
-import org.openrepose.cli.command.common.CommandResult;
+import org.openrepose.cli.command.results.CommandResult;
 
 /**
  *
