+++ project-set/components/cli-utils/src/main/java/org/openrepose/cli/command/datastore/distributed/CacheKeyEncoder.java
@@ -5,7 +5,7 @@ import com.rackspace.papi.service.datastore.encoding.UUIDEncodingProvider;
 import com.rackspace.papi.service.datastore.hash.MD5MessageDigestFactory;
 import java.security.NoSuchAlgorithmException;
 import org.openrepose.cli.command.AbstractCommand;
-import org.openrepose.cli.command.common.CommandResult;
+import org.openrepose.cli.command.results.CommandResult;
 import org.openrepose.cli.command.results.CommandFailure;
 import org.openrepose.cli.command.results.InvalidArguments;
 import org.openrepose.cli.command.results.MessageResult;
