+++ project-set/components/cli-utils/src/main/java/org/openrepose/cli/command/results/CommandResult.java
@@ -1,4 +1,4 @@
-package org.openrepose.cli.command.common;
+package org.openrepose.cli.command.results;
 
 /**
  *
