+++ project-set/components/cli-utils/src/main/java/org/openrepose/cli/command/results/MessageResult.java
@@ -1,7 +1,5 @@
 package org.openrepose.cli.command.results;
 
-import org.openrepose.cli.command.common.CommandResult;
-
 /**
  *
  * @author zinic
